refactor brownie pytest plugin
